Refactor webhook handling for modularity and maintainability
- Centralized configuration management in config/index.js
- Separated middleware for GitHub and Tebex payload verification
- Added IP verification middleware
- Created utility modules for logging, Docker services, and Discord messaging
- Implemented service handlers for Tebex payment events
- Organized routes for GitHub and Tebex webhooks
- Updated app.js for improved middleware and route integration
- Enhanced error handling and logging throughout the application
