+++ .envexample
@@ -1,3 +1,4 @@
+PORT=3000
 WEBHOOK_SECRET=######
 TEBEX_SECRET=########
 DISCORD_WEBHOOK=###/#####-#####
