+++ middleware/index.js
@@ -1,0 +1,4 @@
+// /middleware/index.js
+module.exports.verifyIP = require('./verifyIP');
+module.exports.verifyTebexPayload = require('./verifyTebexPayload');
+module.exports.verifyGitHubPayload = require('./verifyGitHubPayload');
