+++ utils/index.js
@@ -1,0 +1,5 @@
+// /utils/index.js
+module.exports.updateDockerServices = require('./dockerServices').updateDockerServices;
+module.exports.logWithTraceId = require('./logger').logWithTraceId;
+module.exports.sendDiscordMessage = require('./discord').sendDiscordMessage;
+
