+++ utils/logger.js
@@ -1,0 +1,7 @@
+function logWithTraceId(traceId, message) {
+  console.log(`[${new Date().toISOString()}][${traceId}]: ${message}`);
+}
+
+module.exports = {
+  logWithTraceId
+};
